Update caddy_control.py
Add a validate function to caddy control script. It reloads the template of the Caddyfile, executes the setup.sh script to ensure custom certificates will be put into the certificates/temp/ folder, and then validates the caddy configuration with "caddy validate --config ..."

This action ensures that the Caddyfile is actually valid before reloading the service in the GUI, preventing User Error.

The exact Error Message is printed as message. For example:

{"message": "Error: adapting config using caddyfile: ambiguous site definition: example.com:443", "status": "failed"}

and

{"message": "Caddy configuration is valid.", "status": "ok"}