New validate api action + Validation model fix (#98)
* Update Caddy.xml

Fix stray empty line in model

* Update general.xml

Fixed stray whitespace, changed the names of the new Log options and added a little more help text.

* Update setup.sh

Cleaned up the setup script to leverage the -p of mkdir to create full paths.

* Update caddy_control.py

Add a validate function to caddy control script. It reloads the template of the Caddyfile, executes the setup.sh script to ensure custom certificates will be put into the certificates/temp/ folder, and then validates the caddy configuration with "caddy validate --config ..."

This action ensures that the Caddyfile is actually valid before reloading the service in the GUI, preventing User Error.

The exact Error Message is printed as message. For example:

{"message": "Error: adapting config using caddyfile: ambiguous site definition: example.com:443", "status": "failed"}

and

{"message": "Caddy configuration is valid.", "status": "ok"}

* Update actions_caddy.conf

Add new validate action

* Update caddy_control.py

Removed the reload from the validation. Will be handled before the API call in the service controller.

* Update caddy_certs.php

Remove the echo because this script gets triggered quite often.

* Update setup.sh

Comment out the echo because this script gets triggered quite often.

* Update Caddy.php

Fix validation model to use dynamic keys. Fixes issue that validation messages aren't appended to their respective fields.

Based on this:
https://github.com/opnsense/core/commit/1ec272ad6549ddffdbb0c274ed40cb81b56891c7

Thanks AdSchellevis for pointing me into the right direction, very appreciated.

* Update reverse_proxy.volt

Expand Apply button to include an "onPreAction" that checks the "/api/caddy/service/validate" API if the generated Caddyfile is valid.

Only continue and restart the caddy service if the Caddyfile is valid, otherwise show a message with the error message.

* Update general.volt

Add the same validation functionality to apply button in General Settings.

* Update ServiceController.php

Add API endpoint to Service Controller to support the new /api/caddy/service/validate