Update dnsprovider.xml
Change Metaname back, I guess there is an additional validation that checks if the right info is there. The Caddyfile validation catches this. Since I don't know what the right data is, I can't test if the Caddyfile would validate with this. I'm leaving Metaname as it is, maybe somebody uses it in the future and can help.